misc: move logger setup and initialization into separate class
