+++ src/core/main.h
@@ -247,9 +247,7 @@ class ColorButton { };
 class SCNotification { };
 #endif
 
-#define PLOG_OMIT_LOG_DEFINES 
-#define PLOG_NO_DBG_OUT_INSTANCE_ID 1
-#include <plog/Log.h>
+#include "utils/Logger.h"
 
 #ifdef __STANDALONE__
 #include "standalone/inc/atl/atldef.h"
