+++ standalone/cmake/CMakeLists_bgfx-android-arm64-v8a.txt
@@ -269,6 +269,8 @@ add_library(vpinball SHARED
    src/utils/hash.h
    src/utils/objloader.cpp
    src/utils/objloader.h
+   src/utils/Logger.cpp
+   src/utils/Logger.h
 
    third-party/include/imgui/imconfig.h
    third-party/include/imgui/imgui.cpp
