+++ standalone/cmake/CMakeLists_bgfx-ios-arm64.txt
@@ -274,6 +274,8 @@ add_executable(vpinball
    src/utils/hash.h
    src/utils/objloader.cpp
    src/utils/objloader.h
+   src/utils/Logger.cpp
+   src/utils/Logger.h
 
    third-party/include/imgui/imconfig.h
    third-party/include/imgui/imgui.cpp
