+++ standalone/cmake/CMakeLists_gl-macos-arm64.txt
@@ -278,6 +278,8 @@ add_executable(vpinball
    src/utils/hash.h
    src/utils/objloader.cpp
    src/utils/objloader.h
+   src/utils/Logger.cpp
+   src/utils/Logger.h
 
    third-party/include/imgui/imconfig.h
    third-party/include/imgui/imgui.cpp
